put hacky parameters in the channel profile. Added voltage controlled servo demo.

+++ Firmware/R6/Clay_C6_Firmware/Sources/Utilities/PWM_Utils.c
@@ -20,8 +20,8 @@
 ////Local Prototypes///////////////////////////////////////////////
 
 ////Global implementations ////////////////////////////////////////
-uint16_t Scale_Percent_Uint16(uint8_t percent) {
-   return (uint16_t) (((double) percent / 100.0) * 0xFFFF);
+uint16_t Scale_Percent_Uint16(double percent) {
+   return (uint16_t) ((percent / 100.0) * 0xFFFF);
 }
 
 ////Local implementations /////////////////////////////////////////
